Larger Canvas Size Support
* SampleBuffer class to hold the sample buffer and alpha buffer of render as double[][] instead of double[]
* Changed BitmapImage class to use int[][] instead of int[], to be consistant with SampleBuffer
* Made BitmapImage.data protected for better encapsulation.
* Refactor things to work with the new SampleBuffer and BitmapImage. And use the getters and setters instead of directly accessing array (which is faster in most cases too).
* Fixed wrong comment in FloatingPointCompressor class for decompression logic
* PngFileWriter now takes BitmapImage instead of int[]; PfmFileWriter does similar internally.
* BitmapImage function local variables refactored to correct names.
* BitmapImage#vFlipped() now uses System.arraycopy and is thus much faster.

Known Issues:
* Denoiser plugin no longer can directly access double[], and will need to be updated to use the SampleBuffer (and while changing that, may as well switch to the existing PfmFileWriter)
* Limits "Render Preview" canvas size in RenderCanvasFx class to prevent strange JavaFx crashes. This should be changed in the future. The current state can only view the top left corner of the render. I might change this to only show the middle 4k x 4k section, but ideally we should have some zoomed out preview which isnt so big that it crashes (and doesnt take much CPU time away from actual rendering).

+++ chunky/src/java/se/llbit/chunky/renderer/renderdump/PixelConsumer.java
@@ -24,5 +24,5 @@ public interface PixelConsumer {
    * @param g          Green pixel value
    * @param b          Blue pixel value
    */
-  void consume(int pixelIndex, double r, double g, double b);
+  void consume(long pixelIndex, double r, double g, double b);
 }
